Desktop/python_go_complete_game/go_complete_game_python/Deliverables copy/7/7.1/requirements.txt
decorator==4.4.0
future==0.18.1
PyContracts==1.8.12
pyparsing==2.4.2
six==1.12.0